requirements.txt
pytest
Pillow
playwright